wsgi: Don't break HTTP framing during 100-continue handling
Expect: 100-continue is a funny beast -- the client sends it to indicate
that it's willing to wait for an early error, but

- the client has no guarantee that the server supports 100 Continue,
- the server gets no indication of how long the client's willing to wait
  for the go/no-go response, and
- even if it did, the server has no way of knowing that the response it
  *emitted* within that time was actually *received* within that time
- so the client may have started sending the body regardless of what the
  server's done.

As a result, the server only has two options when it *does not* send the
100 Continue response:

- close the socket
- read and discard the request body

Previously, we did neither of these things; as a result, a request body
could be interpreted as a new request. Now, read and discard the body
(in part because we already have tests that exercise pipelined requests
after an early response).

Note that this is some pretty *old* bad behavior; however, prior to
5b5afd1859dd181c14d5dc1be6da1bc1039e52bd we actually did the right
thing!