wsgi: Give applications some say in how much to discard between requests
Some applications may have hard limits for how large a request body can
be and either reject it immediately (because the Content-Length said it
was too big) or read a decent bit, realize the body's too big, and
*then* reject it. This latter case can happen with chunked transfers
where there's no real indication of body size.

In either event, it's on us (as the WSGI server) to discard the body and
look for the next request if the connection should be kept alive.
However, this can lead to a lot of wasted cycles if the client was
trying to send a 100GB payload (or worse, is just piping from
/dev/zero).

Add a new environment key, eventlet.max_discard, that the application
(as sole judge of what a "reasonable" body size is) may set to limit how
much should be read between requests. If more than that many bytes are
read while looking for the end of the body, an IOError will be raised
which eventually closes the connection.