wsgi: Give applications a way to trash a connection
Some applications may have hard limits for how large a request body can
be and either reject it immediately (because the Content-Length said it
was too big) or read a decent bit, realize the body's too big, and
*then* reject it. This latter case can happen with chunked transfers
where there's no real indication of body size.

In either event, it's on us (as the WSGI server) to discard the body and
look for the next request if the connection should be kept alive.
However, this can lead to a lot of wasted cycles if the client was
trying to send, for example, a 100GB payload (or worse, is just piping
from /dev/zero).

Take a page from how websockets communicate back to the protocol layer
and have HttpProtocol.close_connection shim back into WSGI_LOCAL. If an
application wants to prevent the socket from being re-used, it can now
just set wsig.WSGI_LOCAL.close_connection to True.